[mojo] Rework package compilation to store post parser bytecode (#31129)
This PR reworks the way that packages get built to just store the post-parser
bytecode, leaving the library generation phase to be done lazily when the
package is used. When building a package currently, we only store the pre-elaboration
bytecode. The idea at the time was that we would specialize this IR
for a specific target when using the package, but there are a lot of issues
that have popped up because of this:

* The parser wants to have post-parser bytecode to do some interpretation
  and analysis (which currently just fails if that happens).
* Specializing a package for a specific build configuration makes it really
  difficult to switch between different configurations (e.g. debug and release). We
  have special cmake build modes that substantially increase the pain of debugging
  issues (-O0 uses an optimized stdlib -> don't even have stack traces).

This came out as part of resolving #22563, #30035, and #29010.

There are a lot of followup cleans that can happen after this, but wanting to keep this
as tight as possible.

MODULAR_ORIG_COMMIT_REV_ID: 9f317ce724ffc22aec8e169f8b7b6b3b8599bd68