[mojo-stdlib] Run non-GPU tests with `debug-level full` (#31269)
To ensure all the (non-GPU) tests run with debug mode, replace the missing `RUN`
directives that are missing `-debug-level full` to add it.

This also has the benefit of making debugging these tests better using our
debugger and the debugger having a consistent flow for launching/running the
test processes.

Note:
- Skip doing this for the GPU tests since the kernels are not loaded in debug
  mode, so that mismatch will make them not work.

modular-orig-commit: 72243f0fc0a1ece6fb029e2f44190b34fdf1602a