[Mojo] Improve error message when passing type instead of instance (#31405)
Similarly to [Internal link] this patch
improves the error message for a common type/mistake, when a user would
pass a type value instead of an instance, and we would emit something
along the lines of `cannot pass 'Foo' value, parameter expected 'Foo'`.
From now, the error message is clarified to highlight the metatype
nature of the operand in these cases, and also append a hint on how to
fix the problem.

MODULAR_ORIG_COMMIT_REV_ID: 4aa53677686c46558300ba26e34face75b7c1705