[Support] Enable use of client certs for telemetry (#31293)
Eventually, this will be enforcing. It is critical for usage and billing
data, to ensure that the client is reporting for the correct ID and has
been authorized. In the future, the telemetry endpoint will also reject
non-authorized users.

MODULAR_ORIG_COMMIT_REV_ID: dceabde73a788ec97a11f323fcdbbc3bc421e86b