[mojo-stdlib] Mark many types as `Boolable` (#31460)
Many types implement `__bool__` but haven't declared conformance to the
recent `Boolable` trait. So, go ahead and do that across the board.

This requires removing the constructor in `Bool` for any `Boolable`
types and relying on the explicit conversion at the call sites where
needed.

It's not obvious (I think it's a library bug) for why we need the
explicit `.__bool__()` calls in `****.mojo` since the `utils._Optional`
supports `__and__` already.

modular-orig-commit: ad4e7fdd3059120fd80e1aae3a52a0889c450f51