[mojo-stdlib] Move `test_buffer_slice` to `test/NN` (#31582)
`test_buffer_slice` uses things from `NN` in its testing which is purely
a kernels package and not a stdlib thing. Since we won't be
open-sourcing the `NN` package when open source the stdlib, move this
test to `Kernels/test/nn`.

Now, the only remaining uses of `NN` imports are in GPU tests in the
stdlib. We may want to move those in the future so things are layered
properly (e.g. stdlib has no references to `NN` in any of its tests).
Alternatively, we could just not compile those tests in the open source
version so the references aren't unresolved.

Related to [Internal Link]

modular-orig-commit: f7c1fb26b9376a402a156427d88192bd7af8480a