[KGEN] Hash all mojo files when looking up KGEN cache (#31648)
Stacked PRs:
 * __->__#31648
 * #31638
 * #31104

--- --- ---

### [KGEN] Hash all mojo files when looking up KGEN cache

We are pulling the KGEN cache in CI based only on the generated
KGENVersion. This is incorrect, because this only include a hash of the
compiler source code plus the builtins. That means changes to code
outside this set will not cause a new cache to be pulled, potentially
resulting in a cache hit with an old cache version. This explains why
there have been parser caching issues in CI only with non-stdlib
packages.

First, remove `stdlib/builtin` from the KGEN version. The compiler
should not depend on any Mojo source code. Now, include a hash of all
the `*.mojo` files in the repo when pulling the KGEN cache in CI. This
does mean

MODULAR_ORIG_COMMIT_REV_ID: 9158be4f646ef54f53ba09047cad9d6f1b094bf2