[mojo-lang] Lazily resolve decls from source (#31725)
Now that parser caching is out, we can lazily resolve decls from source
code because we don't need the full module to be resolved for caching.
This actually significantly speeds up the mojo parser unit tests
(2-2.5x).

This also requires a hack where we resolve the signatures of children of
a decl parsed from source when a bytecode decl depends on it, because it
requires the symbol to be registered in the symbol table. Previous this
worked because we always fully resolved everything from source.

MODULAR_ORIG_COMMIT_REV_ID: 9cef870d86a99e89a126c18c613045c665c72ca1