[Engine API] Separate Input Spec specified at compile time of torch model from TensorSpec. (#31598)
We were using M_TensorSpec in three places
- One at compile time to specify the valid specs for a Torch model.
- At runtime to specify the tensor spec for execution.
- To inspect model metadata.

The first usage is a bit inconvenient for the user since they also need to specify a name. This name is basically a placeholder and effectively ignored and is not guaranteed to be the same name they need to use at execution time. To make this convenient we introduce a new type to specify compile time valid shapes.

This was the only place where user needed to create TensorSpec with dynamic shapes/dynamic ranks. Now that's done with the new type it is removed from doc string of TensorSpec.

Note: Implementation wise we still use M_TensorSpec under the hood. That's a convenience thing more than anything else and can be changed later without user impact.

All the existing usages in the codebase have been updated along with this change.

While there noticed that we were using `using namspace` declaration in headers. Removed that as well.

MOJO_MAX_ORIG_REV_ID: 9da9f7d8893cb3d1e1523bcb0fb365f0427aa79a