[mojo-stdlib] Make `PythonObject` `SizedRaising` (#31894)
Some types don't implement `len`. So, `len(42)` in Python throws for
example. Match that behavior for Mojo.

Note:
- The test to trigger this properly throws, but then the Python
interpreter get in a weird state and we can't continue in the 
`test_python_object.mojo` file. So, this test is put in a separate
test file for now.

Fixes [Internal Link]

modular-orig-commit: 0c8af0f943b7290cb69f00dbfb4b45c0bb060858