[LLCL] Add withAffinity runtime option (#29032)
This drops the constexpr kUseThreadAffinity option in favor of a boolean that
is actually plumbed through the runtime options. This has separate defaults for
the internal and external (SDK-based) calls, although both are set to true in
order to preserve the current behavior. This may be changed in the future.

MODULAR_ORIG_COMMIT_REV_ID: 0727512c911649dd75a0728a6cdc430d8e23da4b