[LLDB] Reconstruct correctly SIMDTypes and ArrayTypes (#32013)
I made the shameful mistake of considering all array entries in DWARF as
SIMDs, and assuming that other arrays coming from the language would be
represented as structs, which AFAIR happens in some cases.
Doing a correct reconstruction fixes a ton of issues caused by wrong
padding calculations in structs, as the alignment of an ArrayType is not
the same as a SIMDType.

MODULAR_ORIG_COMMIT_REV_ID: 1b00feaa71167eaadd8d04e69c1e57a1abd41c41