[Mojo] Use indices to represent variadicness in `TypeCheckedParamList` (#32046)
This is the first proper step to move the parser to a model where it
represent variadic and pack args/params using indices instead of flags
on a signature. This patch starts to push this change through from top
to bottom: `TypeCheckedParamList` now remembers which parameters were
variadic instead of just storing a flag if any of them were. In a
subsequent patch this information will be pushed into the IR.

MODULAR_ORIG_COMMIT_REV_ID: c76de18d9485f8b5f6085d43556736aa8cb4b829