[API] Constrain number of threads to 16 for DE (#31753)
This uses the MODULAR_RELEASE_PACKAGE_BUILD to constrain the
DE to only run on a max of 16 cores. The constraint
is only enforced if modular was built using the
MODULAR_RELEASE_PACKAGE_BUILD flag.

Note that this needs to be verified that it does
do what it's supposed to do, but I have no idea
how to test this.

Closes #27841
MODULAR_ORIG_COMMIT_REV_ID: 7c2840f8ffc8027203cbf76608c33a2b1698de0c