[LLCL] Simplify limits and default calculations (#31947)
If affinity is not set, then we should not be anchoring on the number of
physical cores, instead we should be using the logical cores. Otherwise, we
risk not using the full capacity of the machine.

This change also simplifies the application of limits and computation of the
affinity set. We can now use affinity effectively even if CPU limits are set.

MODULAR_ORIG_COMMIT_REV_ID: 719e5fee1a96e699d92292b65828007a42a9eeb0