[mojo-lang] Move recursive call checking to LowerSemanticCF (#32153)
Recursive calls are easy to do in Mojo in a few cases, particularly
with implicit conversions.  Previously we had a very narrow check in
the parser that required that all the parameters and arguments be
lexically equal in order to diagnose the problem.  This caught a few
problems but missed many others.

The issue is that we need to be able to understand that conditional
recursion (recursive calls in `if` statements) are ok, but
unconditional ones are not.  This cannot be checked in the parser
because it doesn't have a notion of control flow.

As such, this moves the check from the parser to the LowerSemanticCF
right after semantic CF has been lowered.  This allows us to detect
many more cases.  This fixes #28551 which we now diagnose with:

```
test.🔥:7:25: warning: self recursive call will cause an infinite loop
        self = increment(val)
                        ^
```

MODULAR_ORIG_COMMIT_REV_ID: eccbb9354242a9a7486f7946b926214153bccbab