[Support] Add seed entitlement classes (#32203)
These are the set of entitlements that are currently implied through the
code base. They are not necessarily the final set that we will use in
the product, but can be used to seed use and evolve over time.

MODULAR_ORIG_COMMIT_REV_ID: 1ae677deeaa0a6ee1103d7010d2f9e76cedcdc95