[API] [KGEN] Python: find compilerrt with env var (#32186)
Use modular.cfg's built-in environment variable override from the Python
bindings so that we load the correct compilerrt.
Python packaging expects to load libKGENCompilerRTShared.so from under
site-packages.
This is because packaging patches the wheel libKGENCompilerRTShared.so's
DT_NEEDED to point to hashes versions of its dependency *Globals.so
libraries.
Since only the Python bindings and the installer know where
site-packages is, there needs to be a way for the installer or the
bindings to set this config value.
Here the bindings set KGENCompilerRTShared.so's location with an
environment variable, while issue #32183 tracks tracks a more elegant
solution.

MODULAR_ORIG_COMMIT_REV_ID: 32e264ad5e70533b4377ef12fa86f0777c8d0d26