[KGEN] Don't pull in post-elaboration IR in the elaborator (#32288)
This fixes a flake where the elaborator ends up specializing a generator
from both source IR and precompiled package bytecode at the same time.
The situation is we have

```python
fn another_param_func[a: Int]():
    pass

fn parametric_function[a: Int]():
    another_param_func[a+1]()

fn precompiled_function():
    parametric_function[1]()
```

```python
from some_package import parametric_function, precompiled_function

fn foo():
    alias result = precompiled_function()

fn bar():
    parametric_function[1]()

fn main():
    foo() # 'foo' and 'bar' are elaborated in parallel
    bar()
```

The instantiation of parametric_function[1] from main -> foo is
performed from source, because it is a parametric function whose
definition must be available in the source. The elaborator creates the
kgen.func for parametric_function,a=1 in the IR and kicks of
elaboration of it.

At the same time, the elaborator processes the instantiation of
precompiled_function inside a parameter context. Because the function
can be precompiled, it is pulled out of the precompiled bytecode along
with the same instantiation of parametric_function,a=1 as a transitive
dependency. However, loadElaboratedSymbolsFromBytecode will not insert
a symbol if it already exists, so this does not overwrite the
specialization from source.

Then, the elaborator thinks precompiled_function is complete, and
kicks off the interpreter. The interpreter will then step into a
potentially incomplete specialization of parametric_function,a=1 and
reach a non-concrete call to another_param_func[a+1], causing the
symbol lookup to crash the interpreter.

The original solution was to always pick the function specialized from
the package. However, this was exposing the ABI problem because we
transform the ABI post-elaboration. This PR "fixes" the problem by just
not doing that. See #32285.

Fixes #32198

MODULAR_ORIG_COMMIT_REV_ID: 7ad35c74e09fe255736a30f8cb1db4bee8599e26