[mojo-lang] Simplify copy/move capture logic diagnose errors (#32303)
The `__copy_capture` processing logic was duplicating a lot of lookup
and other logic unnecessarily.  Simplify this, and fix it to check the
cases where the user is trying to capture unknown values.  This was
previously ignored, and we have a lot of bogus captures in the kernel
tree.

While here, speculatively simplify some other name lookup logic. I'm
not sure this will stick but want to give it a try.

MODULAR_ORIG_COMMIT_REV_ID: 00b1b9dd219155e38c7e2a4040084f20f0b964ac