[KGEN] Add mojo-isolated test tool substitutions (#31620)
In other test files, the local variables `tool_dirs` and `tools` are set
in order to define tool substitutions -- meaning, any use of `kgen`, for
example, within a lit `RUN:` command is substituted with the absolute
path to a `kgen` discovered within one of the `tool_dirs`. This helps
ensure that the `kgen` specifically configured and built for this test
run is used, intead of whichever one appears first on the user's PATH.
(A lot of the time, the user's PATH points to the just-built `kgen`, so
this happens to work. But, a user may modify their PATH after
configuring and building this repository, and in that case these tests
still ought to use the correct `kgen`.)

The mojo-isolated tests make use of `kgen-translate`, but they don't
actually define the tool substitution. This patch defines the
substitution to avoid any PATH gotchas in the future.

MODULAR_ORIG_COMMIT_REV_ID: 304a24533a47709acb202616c18acc2df6999af4