[mojo-stdlib] Swap `Variant` to use `!kgen.variant` (#32528)
This is not NFC because it changes the semantics of the type. This PR
changes `Variant` to use `!kgen.variant` as its underlying
representation. This is primarily to fix a bug where `Variant` would
fail inside the interpreter if it contained a pointer value directly.
This is because the interpreter can only keep track of pointer values if
they are stored in memory locations of PointerType. Writing pointer
values into an opaque byte array causes the interpreter to lose track of
it.

This has several other benefits too: now `Variant` will benefit from all
compiler optimizations applied to `!kgen.variant`. This was one of the
original goals of `!kgen.variant`: to be a platform for both error
handling optimization and first-class language variants.

modular-orig-commit: 4260843ae295c905bd16f9a44126ca1e1b041d8b