[REPL] Handle REPLRefResultType in the type system as a regular reference (#32517)
Stacked PRs:
 * #32549
 * #32518
 * __->__#32517

--- --- ---

### [REPL] Handle REPLRefResultType in the type system as a regular
reference

Closes [Internal link]
Closes [Internal link]

In [Internal link] I moved the handling
of REPLRefResultType from the type system into a visualizer. That was
done with the good intention of cleaning up the type system, but ended
up being a bad idea because the variable printing logic in LLDB is not
prepared for such a complex case in which even basic types are managed
by visualizers. This became more visible with
[Internal link] which pretty much broke
printing nested types in the REPL in an almost unrecoverable way.
I did try to fix the issues I found in LLDB, but that code is spaguetti
and damages my soul, so I'm returning to the original approach of
treating REPLRefResultType as a reference type in the type system.

MODULAR_ORIG_COMMIT_REV_ID: be1d401817da92cc6a5aab1a8f8b5709a269d8a6