[Debugger] Fix printing pointer<String> (#32518)
Printing `pointer<String>` causes a crash because of two interesting
events.

- When lldb looks for a formatter for `pointer<String>`, it matches the
pointer with the formatter of `String`. That's how LLDB expects to work.
- Then, our formatter is invoked with the pointer as variable. We try to
access its inner `_buffer` member, but we fail because `_buffer` is not
a child of a pointer, it's instead a child of the `String` struct.

The fix is relatively simple. LLDB expects the typesystem to treat a
pointer as if it was its inner element type when traversing children.
That enables looking for `_buffer` even when the variable is a pointer.

MODULAR_ORIG_COMMIT_REV_ID: 05220abd7af3e5893cb85f84c7af08cf708ef801