[LLDB] Embed decorators in the type name for better data formatter resolution (#32549)
Closes [Internal link]

Few PRs that I have written have brought me so much as this one. In
fact, since I added the `.*` regex to match types with the
`lldb_formatter_wrapping_type` visualizer, I've been fighting LLDB's
value printer system. The reason is that the visualizer based on `.*`
was effectively a sink for any type that wasn't matched with another
visualizer, and if that type didn't contain the decorator, we ended up
using an artificial visualizer that was a passthrough (that's how LLDB
works). This means that all our types were served by visualizers, even
the basic ones. This was hell, because LLDB is full of places with the
following logic:

```
if the type is basic
  do something
else if the type is synthetic (i.e. has visualizer)
  do something else
```

LLDB assumes that synthetic types are not basic, so things start to
break in weird ways. The two most interesting ones are: the first
element of a vector in the REPL shows the type of its inner member;
printing many simple aggregate types was only dumping its address and
not its contents (see the `test_stack_trace.py` changes); and printing
small aggregate types as one liners was not possible.

This PR fixes this by removing `.*` as a visualizer matcher. Now, we are
appending the decorators in the full internal type name, so that we can
do regular regex matching using the decorator name, thus having no more
sinks. This leaves our visualizer system quite clean. LLDB also has a
type name index, so having the decorator name in that index might be
useful as well at some point.

Of course, the ideal solution would be to teach LLDB about decorators,
but that's just infeasible at this point.

MODULAR_ORIG_COMMIT_REV_ID: b948d46ccc576c8b2e911a20d4bfa9486006279d