[mojo][Parser] Set the SMLoc to the op location when loading bytecode (#32565)
This allows for all of the tools and utilities using the location of
the decl to have an accurate location, instead of the location of
the use. This fixes things like go-to-definition for packages in the
Mojo language server.

Fixes #32289
Fixes #24927

MODULAR_ORIG_COMMIT_REV_ID: ad75d4c0ce0f56d68244b796566c0d6e717c0a41