[mojo-lang] Fix memory safety hole with regpassable arguments. (#32636)
In discussion with Jeff about another change, I realized that any
use of `Reference` with generic arguments that can be elaborated
to register_passable type would form a major memory safety hole,
and would generally not work, so I filed Issue #32603.

The problem is that (post-elaboration) arguments with register
passable types are flattened into registers, so any uses of the
address of the nominal arguments will be invalid.

This isn't solvable - you can't flatten an argument to a register
AND have it maintain an address from the caller at runtime.  As such
we need to disallow this edge case.

In practice, doing so is more conceptually difficult than implementation
difficult.  We just notice when this happens and cast the arguments
to a sentinel lifetime type which refuses to convert to a `Reference`.

This fixes #32603: References to borrowed args in generics miscompile
when instantiated on regpassable types

MODULAR_ORIG_COMMIT_REV_ID: 778b59e48469eae2a3d3d55f06f9c90cc00e9e75