[Kernels] remove AVX512 special case path in floor() (#32649)
Remove a special case for SIMD[float32, 16] + AVX512 that would
explicitly use the rndscale.ps intrinsic: the default llvm.floor
intrinsic already does this. The usage of rndscale.ps also didn't mask
FP precision exceptions which the default version did, so this is now
consistent with other variations of floor().

modular-orig-commit: ac2898d52d77deb44ce50a17b1d88ade70e66e5c