[Kernels][Stdlib] Optimize stencil generator and adapt Pool ops. (#32208)
Addresses [Internal link]

Currently, pool ops (`avg_pool`/`max_pool`) pass a load function into
the stencil generator that is different depending on whether there is
padding or not.

In the cases where padding applies (including implicit padding when
`ceil_mode=True`) the `load_fn_no_padding` is currently passed, which
includes an `if` check for the boundary conditions. This is executed for
all points in the padded input array.

This PR addresses this by doing the following:
1) Remove the differentiation between `load_fn` and `load_fn_no_padding`
in the pool kernels themselves and only uses `load_fn` which loads as
needed from the input buffer.
2) Passes the input shape to the stencil generator.
3) Stencil generator does the boundary checks (also using the
input_shape info from (2)) in the `for` loops as in the below; note that
these 4 `max`/`min` checks are evaluated once per output point (vs the
`if` statement which was there before that evaluated 4 `or` cases for
_each_ value in the pool kernel`).

```
                 for i in range(
                    max(0, lower_bound[0]),
                    min(input_height, upper_bound[0]),
                    step_i,
                ):
                    for j in range(
                        max(0, lower_bound[1]),
                        min(input_width, upper_bound[1]),
                        step_j,
                    ):
                        let point_idx = StaticIntTuple[rank](
                            indices[0], i, j, indices[3]
                        )
                        let val = load_fn[simd_width, type](point_idx)
                        result = compute_fn[simd_width](point_idx, result, val)
```
4) Originally I had non-overlapping loops for the padded areas loading
the pad value. However, I realized that the padded area (which would
have negative values for `lower_bound[]`/`upper_bound[]` does not need
to be initialized with the pad_value (which was taking place in
`load_fn` function in the pool kernels before).
Reason is that `result` is initialized by `compute_init_fn()` to the
appropriate value (e.g.,`neginf()` or `0` in the `max_pool`/`avg_pool
cases`).
Loading and calculating the result for boundary locations, which would
be:
```
                let val = pad_value
                result = compute_fn[simd_width](point_idx, result, val)
```
would therefore make no difference to not doing it at all (comparing 0
to 0 would yield 0, for example).
NN_ORIG_REV_ID: e7e661eeed8d7c97e1c8beb6d4498a9c91c81fb6