[MAX] Add a register.op function that would enable op registration (#32765)
This adds a public decorator that would allow folks to extend the engine
with a custom op. Note that this is not sufficient to actually use the
extensibility API since the entire MOGG package is not fully part of the
stdlib. To be able to register and use an op, we need to move a bunch of
files from MoggExperimental to a place in the stdlib that allows folks
to use them (e.g. `empty_like`) and the `MoggTensor`.

Closes [Internal link]

MOJO_MAX_ORIG_REV_ID: b180a78d0e3b29b76e2f345ce9e0e30bcb58a42c