[mojo-lang] Move return slot to the end of the argument list. (#32790)
This patch changes the mojo parser to add the byref-result slot to the
end
of the argument list instead of the beginning. This has the advantage of
following source order, but also makes it stop messing around with the
"self argument is argument 0" invariant, which simplifies a bunch of
code.

This fixes #30134

MODULAR_ORIG_COMMIT_REV_ID: 6a93b804506e1a7ceca173f04198a9d2364740c6