[Debugger] Ensure a response is received when using mojo debug --rpc (#32737)
This makes `mojo debug --rpc` read a response from the RPC server and
display error messages when applicable. Otherwise, there's no way for
the client to know if the request failed or not. This also ensures that
a correct exit code is returned.

I had to change a bit how the request is read by the the RPC server,
because it was only processing the request after the connection was
closed, which means that there was no way for the client to receive a
response.

MODULAR_ORIG_COMMIT_REV_ID: 751925f7a2d96b8a9700f5b93dd193acead5b1dd