[KGEN] fold VariadicSplat->VariadicSize (#32913)
If the source of a VariadicSize op is a VariadicSplat op, then the size
is known to be the NumElements from the splat op.

This came up because I was looking at the LLVM code for something like:
`SIMD[type, 16](Scalar[type](value))`. This is handled through the SIMD
constructor that takes a variadic list. There is a shortcut for
len(list)==1, but this wasn't simplifying until LLVM codegen. With the
extra folder, the LLVM input is already reduced.

MODULAR_ORIG_COMMIT_REV_ID: ddc0a735fdb6235e6f96c0cfd9c5336681e74695