[mojo-lsp] Fix go-to-definition of modules (#32942)
This issue is caused by the fact that we have been using 0 for the
line/column of modules, which is treated as an invalid location (the
expectation is 1-based line/column). This change makes the line/column
of modules 1 instead, which still covers the desired "points to the
beginning of the file" behavior.

Fixes #32919

MODULAR_ORIG_COMMIT_REV_ID: 2adf4195ba1dfead701ef9a8656b7ab3bf750956