[KGEN] add float_literal.convert (#30638)
IE an op to convert from float_literal to MLIR float types.

This op has some hairy logic in the C++ implementation. There is some testing
in this commit of the MLIR folder, but there is more testing in Mojo in a later
commit in this stack because it's easier to test at the Mojo level.

This is part of the addition of FloatLiteral outlined here:

[Internal link]

MODULAR_ORIG_COMMIT_REV_ID: b87deca18f739b07dd9cd262ad227e861c832f97