[mojo-lang] add infinite precision FloatLiteral (#30640)
The design doc for this work is here:

[Internal Link]

But the main important point for this commit is that we are adding
FloatLiterals that are actually infinite precision rational numbers.

Every literal that we can write with our float syntax is a rational
number (IE float literals use base-10 notation with an optional base 10
exponent). However, many numbers that we can write as literals don't
have a finite representation in binary floating point, such as `0.2`,
which in theory results in an infinitely repeating bit string. Since we
don't actually have infinite memory, this means that if we use an actual
floating point representation, for every `0.2` or similar literal, we
have to make an up-front decision about the bit length. There is a
tradeoff here between using more memory to have a more precise
representation and saving space because more precision is unnecessary.
Using rational numbers sidesteps this.

Rational numbers have issues of their own. For example, many float
numbers, especially with large exponents, have a smaller in-memory
representation as infinite precision floats than as infinite precision
rational numbers. I believe that for the typical number range that we
will see in practice, the sizes are not different enough to matter much
when accurate float strings are finite, but rationals are much better
for the case of floats that require infinite bitstrings.

modular-orig-commit: 539a40578d03611e23d97695bccd067be16604f9