[KGEN][Interp] Fix incorrect 0 write when externalizing pointer region (#33032)
When externalizing a pointer region, the interpreter was writing 0 to
the memory location to "canonicalize" the memory. This was mainly done
for two reasons: if for whatever reason the address leaked without
correct lowering, the user will get a nullptr exception at runtime (not
sure if this is the best thing). The second is to just make caching
better, so as to not expose the address workings on the interpreter.

The bug was that the 0 was not written to the right offset. Just remove
the write because it isn't useful, and having the pointer values written
into the MLIR resources helps debug.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 3584a928e253fbe4c6edfed91e66bcf61e143c56