[mojo-stdlib] Use `vector.__get_ref` in `Dict` (#33122)
As part of the eventual removal of `__getitem__` from `DynamicVector`,
this explores moving `Dict` over to always using the explicit
`__get_ref` function. It's not trivial to fix all of the callers of
`__getitem__` over in one mega patch. Migrating some bigger spots over
to `__get_ref` and then over to `__refitem__` may be the right
incremental play; we'll see.

modular-orig-commit: dcf207f2c78b32b056bdbd099a9fac25ed597b66