[mojo-stdlib] Add `__ne__` requirement to `EqualityComparable` (#33160)
Since we don't currently synthesize `__ne__` from `__eq__`, add `__ne__`
function as a requirement for conformance to the `EqualityComparable`
trait. This will allow us to make easier use of `EqualityComparable` by
assuming existence of `!=` instead of doing the rewrite/assuming it is
equivalent to `not(x == y)`. This will get used in our testing framework
functions like `assert_equal` and `assert_not_equal` for example in a
follow-up.

This exposed the fact that `Set` had conformance to
`EqualityComparable`, but was missing an `__ne__` function. Add that
function and adjust the tests to use it.

Related to [Internal link]

LAYOUT_ORIG_REV_ID: 22fe7cdb5c6d0a152c0666f045992ff487767398