[Bazel] GenericML, KGEN, Kernels overlay pass. (#30859)
This PR enables a bunch of functionality (but not full) in `GenericML`,
`Kernels`, and most importantly `KGEN`. This patch will fully build
`//KGEN:Compiler` and several other main targets.

`//GenericML:Support` is close, but requires additional passes in `KGEN`
for `CompilerRT`. Given the size of this patch, I think it is best to do
a courtesy flush and allow others to tag on fixes/rules as we go.

---------

Co-authored-by: abdul dakkak <adakkak@modular.com>
Co-authored-by: Scott Main <scott@modular.com>
Co-authored-by: Zac Bowling <zac@modular.com>
Co-authored-by: Ian Tramble <117689425+itramble@users.noreply.github.com>
Co-authored-by: Patrick Dougherty <patrick.dougherty@modular.com>
Co-authored-by: Hengjie Wang <86926839+hengjiew@users.noreply.github.com>
Co-authored-by: Arjun Surendran <arjun@modular.com>
MODULAR_ORIG_COMMIT_REV_ID: 2f2d6628957f617ad671412cfa42aa45f2ee8dd3