[mojo-stdlib] Add `EqualityComparable` trait for most types (#33269)
Declare conformance for most types that implement `__eq__` already. Fill
in the gaps for a few types that were missing implementation for
`__ne__`.

Worth calling out:
- `SIMD` can't declare conformance to `EqualityComparable` since the
current semantics of its `__eq__` are to return a `SIMD` vector of
`Bool`s which breaks simple generic programming mental models over
traits. I recommend we change that in a follow-up, but it's worth doing
in a separate PR.
- `PythonObject` dunder methods all return a `PythonObject` (and raise)
rather than return a normal primitive type like `Bool`. So, don't try to
do anything with that right now.

Fixes [Internal Link]

modular-orig-commit: 7f547d635ed6384333494b914a49065772858ac2