[KGEN] Teach the reducer to stub functions out one at a time (#33275)
This is slow but it performs a linear search for all the functions that
can be stubbed out to reproduce the same failure. This does not search
the order in which the functions are removed, as it assumes that doesn't
matter for reaching a minimal-ish example.

MODULAR_ORIG_COMMIT_REV_ID: a73e286c127b01abdac680e42c13eaf8e7d06f65