[TensorFlow] [GenericML] Canonicalize TensorFlow SavedModel paths (#33106)
This was originally for #16934, until I realized that we already don't
go into the TensorFlow driver any more for the code described in the
issue. This means the errors generated by this canonicalization logic
will not ever actually make their way to a user today, but it does make
it more robust in case they ever do. Furthermore, this adds a
long-requested feature of stripping `saved_model.pb` off the end of a
model path if provided by a user, if that would otherwise make the model
path valid for TensorFlow.

MODULAR_ORIG_COMMIT_REV_ID: 92dd3903bbb87b5393ff7c894c8c4a99832bbd4d