[Debugger] Improve the debug me function (#33408)
- Rename the method to `attachToNewRemoteDebugSession` which indicates
that a brand new debug session will be launched and that the remote
debugger doesn't need to exist as an instance.
- Extract some code as `waitForDebuggerToAttach`, which can be used in a
standalone manner, for whoever does debugging not from VSCode.
- Change the behavior of `attachToNewRemoteDebugSession` to always end
with `waitForDebuggerToAttach()`, so that it will always suspend itself,
given the user the chance to attach manually.

MODULAR_ORIG_COMMIT_REV_ID: 77ef347750424aaaf41750291b915f81a851e670