[mojo-lang] Narrow a hack in SubscriptExpr emission. (#33541)
We currently allow direct subscripting on values of `!kgen.variadic`
type.  This is actually used for type packs, e.g. in variant.mojo
because of how we represent variadics in parameters. This is all yuck,
but I don't think we need the dynamic form of this, so trim this down.

MODULAR_ORIG_COMMIT_REV_ID: cf86d024a623bc3160825d805ef753aca41061de