[mojo-lang] Fix off-by-one issue emitting variadic parameter calls (#33618)
How's this for niche: in variadic calls to `__init__` methods
in a parameter context, when the variadic list has a memory type
on the callee but register type for the caller (due to traits),
we would fail to box into memory, causing horrible elaborator
failure.

The root cause is we didn't account for the init_self argument
index. This fixes #33579

MODULAR_ORIG_COMMIT_REV_ID: c17dc7944f535337efb4a6ab109b7afd0ed2e9cf