[Support] Remove Cryptography/Symmetric (#33511)
Since this library is not currently being used, it is best to delete
rather than provide temptation. After sleuthing, it seems the original
motivation was to provide encryption for customer models and the CAS
[1,2], however:
* For customer models we are likely to pursue a much simpler and more
reliable process & policy. (Having encryption would require such a
policy for access to the keys anyways, so this effectively saves a
step.)
* For the CAS, we're a long ways off iterating on features such as
encryption. It is more likely that we will leverage encryption built-in
to underlying storage mechanisms (e.g. customer-owned keys for buckets,
etc.).

In the future, this functionality can be resurrected if needed.

[1] [Internal link]
[2] [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 5122239150ff28dd62f5b4b1c5f7ece1e4d4d62e