[KGEN] Teach ResolveCompilerPromises to handle recursion (#33687)
This is a slight rewrite/rearranging of the ResolveCompilerPromises
algorithm from a simple bottom-up DAG traversal to handle recursion.

This was achieved by separating the algorithm into an analysis phase and
a rewrite phase. The analysis phase can be fixed-point iterated:
analysis state is stored in temporary data structures on the operations
of interest or in the analysis state. Each iteration of the analysis on
a node computes a set of required promises for each function. If a run
adds new promises to the set, then the analysis is signalled to change.
Then, the rewrite phase uses the computed results to lower and rewrite
any relevant operations.

Fixed-point iteration is bound within an SCC. The pass first computes
the SCCs from a virtual root node to all nodes in the callgraph (formed
by special capturing edges), then a worklist resolves the DAG of SCCs by
fixed-point iterating the algorithm within each SCC.

The next PR will parallelize the pass over the SCC worklist.

Fixes [Internal link]

MODULAR_ORIG_COMMIT_REV_ID: 1c2dd0249b0919e47f40d05e25a1371c22323cd6