[GPU][Profiling] Plumb nvtx-3 and include basic marker POC (#33373)
This PR finishes integrating nvtx-3 with the `profiling_nsight` build
profile (i.e. `cmake-modular-nsight-profiling`), such that nvtx-v3 calls
can generally be made in the code. There is one *very* minimal example
of such a call, which shouldn't be taken to be more than a POC/demo.

Version 3 of nvtx is header-only. I tried to follow the same process
that was used for the cuda runtime headers (and other third party deps).
Given that we are just beginning to scratch the surface when it comes to
GPU profiling, I think it makes sense to include access to all of nvtx.
We can always thin the interface in the future (something something
premature optimization).

In principle, the included synchronization of the option name with the
flag in `build.py` should have enabled the `profileStart()` and
`profileStop()` filters in `ExecutionDriver`, and I have verified that
they are called. However, the profile in nsight still includes the
entire execution timeline. I'm not sure why.

## Future work

I've been trying to use process ranges to capture context from the LLCL
profiler. This is complicated by the fact that the target ranges for GPU
executions don't form a stack (due to async ops), so we need to track an
nvtx-generated event id that doesn't currently have anywhere to live in
the LLCL profiler. I'm working on it, but didn't want to block this work
on that.

Another issue with GPU profiling is the aforementioned async nature of
the GPU ops. The chrome-trace event types that we are using expects the
events to form a stack, which causes
[Internal link] There is a different
event type to handle asyncs that should help if we migrate to them.

MODULAR_ORIG_COMMIT_REV_ID: b3f18073be48639cff5baaf86960241f1f68be06