[KGEN] Include `modular.cfg` path when `orc_rt` can't be found (#33849)
In certain cases, such as when `MODULAR_HOME` env variable is not set,
we get a runtime error when we try to `dlopen` the `orc_rt`. On
failure/when the library can't be found, also print the specific path to
the `modular.cfg` file that it was trying to read the path to `orc_rt`.

This is a QoI improvement for downstream clients such as those using
`nightly/mojo` compiler since there are incidentally, currently, two
`modular.cfg` files laying around in `MODULAR_HOME` which was confusing
to debug today.

MODULAR_ORIG_COMMIT_REV_ID: e657c7c3d7357a31db4c4c00d7261cdc7ebfdffa