[KGEN] Fix parser-generated debug scopes for closures (#33769)
Today, closure instances are accidentally generated with the debug scope
of the parent function. It should actually take into account the parent
lexical block if it's inside a lexical block, otherwise it will generate
code that seemingly belongs to the outside of the block when it's
inside.

Example:
```python
fn foo(x: Bool):
  if (x):
    fn inner():
      pass
    inner()
  else:
    pass
```

In this case the glue code that calls `inner` should be created with the
then-block's scope, not the function foo's scope.

MODULAR_ORIG_COMMIT_REV_ID: e3835e26388df66adda96dfade22936d5b26546f