[Driver] Make failure inserting to cache not an error. (#32907)
Add a new argument to CacheTransform for callers to decide if failure to write to cache is an error or not. Use this in MEF caching to ignore error if cache write fails at the cost of recompilation next time. If the transformation fails that's an error as before.

MODULAR_ORIG_COMMIT_REV_ID: a876fc3ec7e29cd19333975f53aba05f5cb35fcc