[VSCode] Cleanup the SDK resolution logic (#33899)
With my recently gained knowledge, I decided to cleanup how we resolve
the SDK and cache its results. This effectively allows for modulers to
have LSP and debugger support for files that are not part of the modular
folder regardless of the number of open workspaces, because we've been
setting the modular home path at the modular folder-level when doing
`vscode-init`. This also fixes an issue in which you couldn't use `Mojo:
Attach to Process` if you had two workspace folders open and the SDK was
not configured at the user-level. VSCode!!!!!!

- Add a `Mojo: Install and Configure the Mojo SDK` command that triggers
the same dialog we use when notifying that the SDK is not installed. Now
this suggests modifying the `modularHomePath` setting at the user-level
and not workspace-level.
- Added some logic that allows reusing the SDK config of any active
workspace whenever the resolver fails at doing that for a particular
context.
- Reset the LSP clients, the `file -> LSP` cache, and the `workspace ->
SDK config` cache whenever there's a change to the list of open
workspaces. This is needed because reading the `mojo.modularHomePath`
setting actually depends on the number of active workspaces. I left a
comment in the code mentioning that.
- Make the `file -> LSP` and `workspace -> SDK config` caches correct
under concurrency, i.e. storing a Promise in the cache instead of its
result.
- Move the file watching logic from the LSP-level to the SDK
config-level, so that we can watch more files belonging to the SDK that
can help with restarting the extension properly.
- Updated the documentation and changelog.

MODULAR_ORIG_COMMIT_REV_ID: a55de697df1ec8fb490f97d7286a5d255f4d72ee